# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API EliteRT",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pengeluaran": {
            "get": {
                "tags": [
                    "Pengeluaran"
                ],
                "summary": "Menampilkan semua data pengeluaran",
                "operationId": "17438ca5e742e916663eb39dd2d2e8a1",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data pengeluaran"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pengeluaran"
                ],
                "summary": "Menambahkan data pengeluaran baru",
                "operationId": "77987b3680f32a7ed4cea1943fbba449",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_pengeluaran",
                                    "kategori",
                                    "jumlah_pengeluaran"
                                ],
                                "properties": {
                                    "nama_pengeluaran": {
                                        "type": "string"
                                    },
                                    "kategori": {
                                        "type": "string"
                                    },
                                    "jumlah_pengeluaran": {
                                        "type": "number"
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pengeluaran berhasil ditambahkan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/pengeluaran/{id}": {
            "get": {
                "tags": [
                    "Pengeluaran"
                ],
                "summary": "Menampilkan data pengeluaran berdasarkan ID",
                "operationId": "af0e37dd7c22dc374d90f99e95af7da6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail pengeluaran berhasil diambil"
                    },
                    "404": {
                        "description": "Pengeluaran tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pengeluaran"
                ],
                "summary": "Mengupdate data pengeluaran",
                "operationId": "6148fb32c49f8cf8cb78eec37def638f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_pengeluaran",
                                    "kategori",
                                    "jumlah_pengeluaran"
                                ],
                                "properties": {
                                    "nama_pengeluaran": {
                                        "type": "string"
                                    },
                                    "kategori": {
                                        "type": "string"
                                    },
                                    "jumlah_pengeluaran": {
                                        "type": "number"
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pengeluaran berhasil diperbarui"
                    },
                    "404": {
                        "description": "Pengeluaran tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pengeluaran"
                ],
                "summary": "Soft delete data pengeluaran",
                "operationId": "2d6ca27f2dea0257d0376b06d8189b2f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pengeluaran berhasil dihapus"
                    },
                    "404": {
                        "description": "Pengeluaran tidak ditemukan"
                    }
                }
            }
        },
...
        "/api/riwayat-penghuni/{id}": {
            "put": {
                "tags": [
                    "Riwayat Penghuni Rumah"
                ],
                "summary": "Update data riwayat penghuni rumah",
                "operationId": "f94594cc2502d0ea5f13662c179dee91",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID riwayat penghuni",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tanggal_keluar"
                                ],
                                "properties": {
                                    "tanggal_keluar": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-07-15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Berhasil mengupdate data"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Riwayat Penghuni Rumah"
                ],
                "summary": "Menghapus (soft delete) data riwayat penghuni rumah",
                "operationId": "be39c95a42db2fbd48c2cd1f9026dcb3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil dihapus"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
...
    },
...
}